Add some custom styles for the journal
These additions included a reference style for my APA 6th Edition
references in each post and on pages.
Also included some custom styles for the page on the journal called
“Grid” which is about my custom grid. I named the grid “Gridlandish”,
mostly because it is the polar opposite of anything outlandish; it is
just a super simple grid! I wanted to share the code on my journal so
it was all together with some custom demos that I created using iframes
and resizable divs.